+++ src/components/pages/rncs/Observations.vue
