+++ src/components/pages/rncs/RegistrationDetails.vue
